requirements.txt
# requirements.txt 内容
flask==2.0.1
torch==1.10.0
torchvision==0.11.1
pillow==8.4.0